raspberry-pi/requirements.txt
pyserial
websockets
asyncio
websockets